Repository

+++ src/main/java/com/pranitpatil/latestPrice/repository/InstrumentRepository.java
@@ -5,12 +5,12 @@ import com.pranitpatil.latestPrice.dto.Instrument;
 public interface InstrumentRepository {
 
     /**
-     * This method finds the instrument from repository by id, if not found then @throws {@link com.pranitpatil.latestPrice.exception.NotFoundException}
+     * This method finds the instrument from repository by id.
      * @param id
      * @return Instrument
-     * @throws com.pranitpatil.latestPrice.exception.NotFoundException
+     * @throws com.pranitpatil.latestPrice.exception.NotFoundException if id is not found
      */
-    Instrument getInstrumentById(int id);
+    Instrument getInstrumentById(long id);
 
     /**
      * Saves or updates the provided instrument in repository.
